rewrite existing bitwise operations

+++ heat/core/indexing.py
@@ -121,7 +121,7 @@ def where(cond, x=None, y=None):
         y, (dndarray.DNDarray, int, float)
     ):
         cond = types.float(cond)
-        return (cond == 0) * y + cond * x
+        return types.float(cond == 0) * y + cond * x
     elif x is None and y is None:
         return nonzero(cond)
     else:
